Added menu configuration
